+++ app/src/navigation/misc.tsx
@@ -4,15 +4,15 @@ import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 import React from 'react';
 import { StatusBar } from 'react-native';
 
+import LaunchScreen from '../screens/misc/LaunchScreen';
 import LoadingScreen from '../screens/misc/LoadingScreen';
 import ModalScreen from '../screens/misc/ModalScreen';
-import NewLaunchScreen from '../screens/misc/NewLaunchScreen';
 import SplashScreen from '../screens/misc/SplashScreen';
 import { black } from '../utils/colors';
 
 const miscScreens = {
   Launch: {
-    screen: NewLaunchScreen,
+    screen: LaunchScreen,
     options: {
       header: () => (
         <StatusBar barStyle="light-content" backgroundColor={black} />
